pyproject.toml
[project]
name = "app-reviews"
version = "0.1.0"
description = "Mobile apps reviews analyzer"
readme = "README.md"
requires-python = "<3.12,>=3.11"
dependencies = [
    "app-store-scraper>=0.3.5",
    "fastapi>=0.115.11",
    "google-play-scraper>=1.2.7",
    "jinja2>=3.1.6",
    "langchain>=0.3.20",
    "langchain-anthropic>=0.3.9",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]
